Crate hamming_rs

Modules

lib_avx2

avx2 target specific functions

utils

export utilies (such as cross platform aligned_alloc)

Functions

distance

Computes hamming distance
Assumes x and y have same memory alignment
Uses highly optimized avx2 version if available
fallback on distance_naive if x and y have different alignment or if avx2 features are not available

distance_naive

Computes hamming distance (naive version) copied from hamming bitwise fast crate as author agreed

weight

Computes hamming weight
Uses highly optimized avx2 version if available

weight_naive

computes hamming weight (naive version)